refactor: 테스트 코드 패키지명 컨벤션에 맞게 수정
