+++ src/test/java/gift/repository/MemberRepositoryTest.java
@@ -1,7 +1,6 @@
-package gift.RepositoryTest;
+package gift.repository;
 
 import gift.entity.Member;
-import gift.repository.MemberRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
