+++ src/test/java/gift/repository/WishListRepositoryTest.java
@@ -1,7 +1,6 @@
-package gift.RepositoryTest;
+package gift.repository;
 
 import gift.entity.WishList;
-import gift.repository.WishListRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
